requirements.txt
fastapi
uvicorn
sqlmodel
pydantic
python-dotenv
qrcode
Pillow
